+++ CHANGELOG.md
@@ -3,7 +3,29 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.40.0][0.40.0] - 2019-10-15
+## [0.50.0][0.50.0] - 2020-05-2020
+1. Support Bot API v4.8: (by @danielperez9430)
+   * Add methods: *sendDice()*
+1. Support Bot API v4.7: (by @danielperez9430)
+   * Add methods: *getMyCommands()*,*setMyCommands()*
+1. Support Bot API v4.5: (by @danielperez9430)
+   * Add methods: *setChatAdministratorCustomTitle()*
+1. Support Bot API v4.4: (by @danielperez9430)
+   * Add methods: *setChatPermissions()*
+1. Support for poll_answer (by @JieJiSS)
+
+Changed: (by @danielperez9430)
+ * New message type: *dice*
+ * Fix Bugs in tests
+
+New Test: (by @danielperez9430)
+   * sendDice
+   * getMyCommands
+   * setMyCommands
+   * setChatAdministratorCustomTitle
+   * setChatPermissions
+
+## [0.40.0][0.40.0] - 2019-05-04
 
 Added: